12-11-1037 Roundup Commit
Updating to latest, no real changes

+++ app/src/main/java/com/example/steven/ibeaconmuseum/LocationClasses/GridPoint.java
@@ -6,10 +6,11 @@ package com.example.steven.ibeaconmuseum.LocationClasses;
 
 public class GridPoint {
 
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public GridPoint(int x, int y){
+    public GridPoint(double x, double y){
+
         this.x = x;
         this.y = y;
     }
